+++ client/src/lib.rs
@@ -13,9 +13,10 @@ use solana_client::rpc_client::RpcClient;
 use solana_client::rpc_config::{RpcTransactionLogsConfig, RpcTransactionLogsFilter};
 use solana_client::rpc_response::{Response as RpcResponse, RpcLogsResponse};
 use solana_sdk::commitment_config::CommitmentConfig;
-use solana_sdk::signature::{Keypair, Signature, Signer};
+use solana_sdk::signature::{Signature, Signer};
 use solana_sdk::transaction::Transaction;
 use std::convert::Into;
+use std::rc::Rc;
 use thiserror::Error;
 
 pub use anchor_lang;
@@ -36,7 +37,7 @@ pub struct Client {
 }
 
 impl Client {
-    pub fn new(cluster: Cluster, payer: Keypair) -> Self {
+    pub fn new(cluster: Cluster, payer: Rc<dyn Signer>) -> Self {
         Self {
             cfg: Config {
                 cluster,
@@ -46,7 +47,11 @@ impl Client {
         }
     }
 
-    pub fn new_with_options(cluster: Cluster, payer: Keypair, options: CommitmentConfig) -> Self {
+    pub fn new_with_options(
+        cluster: Cluster,
+        payer: Rc<dyn Signer>,
+        options: CommitmentConfig,
+    ) -> Self {
         Self {
             cfg: Config {
                 cluster,
@@ -62,7 +67,7 @@ impl Client {
             cfg: Config {
                 cluster: self.cfg.cluster.clone(),
                 options: self.cfg.options,
-                payer: Keypair::from_bytes(&self.cfg.payer.to_bytes()).unwrap(),
+                payer: self.cfg.payer.clone(),
             },
         }
     }
@@ -72,7 +77,7 @@ impl Client {
 #[derive(Debug)]
 struct Config {
     cluster: Cluster,
-    payer: Keypair,
+    payer: Rc<dyn Signer>,
     options: Option<CommitmentConfig>,
 }
 
@@ -93,7 +98,7 @@ impl Program {
         RequestBuilder::from(
             self.program_id,
             self.cfg.cluster.url(),
-            Keypair::from_bytes(&self.cfg.payer.to_bytes()).unwrap(),
+            self.cfg.payer.clone(),
             self.cfg.options,
             RequestNamespace::Global,
         )
@@ -104,7 +109,7 @@ impl Program {
         RequestBuilder::from(
             self.program_id,
             self.cfg.cluster.url(),
-            Keypair::from_bytes(&self.cfg.payer.to_bytes()).unwrap(),
+            self.cfg.payer.clone(),
             self.cfg.options,
             RequestNamespace::State { new: false },
         )
@@ -323,7 +328,7 @@ pub struct RequestBuilder<'a> {
     accounts: Vec<AccountMeta>,
     options: CommitmentConfig,
     instructions: Vec<Instruction>,
-    payer: Keypair,
+    payer: Rc<dyn Signer>,
     // Serialized instruction data for the target RPC.
     instruction_data: Option<Vec<u8>>,
     signers: Vec<&'a dyn Signer>,
@@ -345,7 +350,7 @@ impl<'a> RequestBuilder<'a> {
     pub fn from(
         program_id: Pubkey,
         cluster: &str,
-        payer: Keypair,
+        payer: Rc<dyn Signer>,
         options: Option<CommitmentConfig>,
         namespace: RequestNamespace,
     ) -> Self {
@@ -362,37 +367,44 @@ impl<'a> RequestBuilder<'a> {
         }
     }
 
-    pub fn payer(mut self, payer: Keypair) -> Self {
+    #[must_use]
+    pub fn payer(mut self, payer: Rc<dyn Signer>) -> Self {
         self.payer = payer;
         self
     }
 
+    #[must_use]
     pub fn cluster(mut self, url: &str) -> Self {
         self.cluster = url.to_string();
         self
     }
 
+    #[must_use]
     pub fn instruction(mut self, ix: Instruction) -> Self {
         self.instructions.push(ix);
         self
     }
 
+    #[must_use]
     pub fn program(mut self, program_id: Pubkey) -> Self {
         self.program_id = program_id;
         self
     }
 
+    #[must_use]
     pub fn accounts(mut self, accounts: impl ToAccountMetas) -> Self {
         let mut metas = accounts.to_account_metas(None);
         self.accounts.append(&mut metas);
         self
     }
 
+    #[must_use]
     pub fn options(mut self, options: CommitmentConfig) -> Self {
         self.options = options;
         self
     }
 
+    #[must_use]
     pub fn args(mut self, args: impl InstructionData) -> Self {
         self.instruction_data = Some(args.data());
         self
@@ -400,6 +412,7 @@ impl<'a> RequestBuilder<'a> {
 
     /// Invokes the `#[state]`'s `new` constructor.
     #[allow(clippy::wrong_self_convention)]
+    #[must_use]
     pub fn new(mut self, args: impl InstructionData) -> Self {
         assert!(self.namespace == RequestNamespace::State { new: false });
         self.namespace = RequestNamespace::State { new: true };
@@ -407,6 +420,7 @@ impl<'a> RequestBuilder<'a> {
         self
     }
 
+    #[must_use]
     pub fn signer(mut self, signer: &'a dyn Signer) -> Self {
         self.signers.push(signer);
         self
@@ -453,7 +467,7 @@ impl<'a> RequestBuilder<'a> {
         let instructions = self.instructions()?;
 
         let mut signers = self.signers;
-        signers.push(&self.payer);
+        signers.push(&*self.payer);
 
         let rpc_client = RpcClient::new_with_commitment(self.cluster, self.options);